+++ apps/demo/src/app/viewer/viewer.component.ts
@@ -3,9 +3,9 @@ import { ActivatedRoute, Router } from '@angular/router';
 import {
   MimeViewerConfig,
   MimeViewerMode,
+  RecognizedTextMode,
 } from '@nationallibraryofnorway/ngx-mime';
 import { Subscription } from 'rxjs';
-import { RecognizedTextMode } from '../../../../../libs/ngx-mime/src/lib/core/models/recognized-text-mode';
 import { ManifestService } from './../core/manifest-service/manifest.service';
 
 @Component({
