+++ libs/ngx-mime/src/index.ts
@@ -4,6 +4,7 @@ export {
   MimeViewerIntlNoNb,
 } from './lib/core/intl';
 export { MimeViewerConfig } from './lib/core/mime-viewer-config';
+export { RecognizedTextMode } from './lib/core/models';
 export { Manifest as MimeManifest } from './lib/core/models/manifest';
 export { ViewerMode as MimeViewerMode } from './lib/core/models/viewer-mode';
 export { MimeModule } from './lib/ngx-mime.module';
