+++ libs/ngx-mime/src/lib/core/models/index.ts
@@ -1,0 +1,2 @@
+export { RecognizedTextMode } from './recognized-text-mode';
+export { ViewerMode } from './viewer-mode';
